pyproject.toml
[project]
name = "genericmiot-vacuum-room-cleaner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click==8.0.4",
    "homeassistant>=2024.12.5",
    "python-miio>=0.6.0.dev0",
]
